+++ Keyboard_Avoidance/Prototypes/WireframeViewController.swift
@@ -31,7 +31,7 @@ class WireframeViewController: UIViewController, UITextFieldDelegate {
         view.addSubview(wireframe)
         
         // Use this to ignore bottom safe area (in `keyboardLayoutGuide` as well).
-        view.insetsLayoutMarginsFromSafeArea = false
+        // view.insetsLayoutMarginsFromSafeArea = false
         
         // Constraints.
         wireframe.translatesAutoresizingMaskIntoConstraints = false