+++ Keyboard_Avoidance/ViewController.swift
@@ -32,8 +32,7 @@ class ViewController: UIViewController {
     lazy var content = UIStackView()
         .vertical(spacing: UI.spacing)
         .views(
-            HeaderView()
-                .withFixedHeight,
+            HeaderView(),
             emailTextField,
             givenNameTextField,
             familyNameTextField,
@@ -61,12 +60,20 @@ class ViewController: UIViewController {
         // Hierarchy.
         view.addSubview(body)
         body.translatesAutoresizingMaskIntoConstraints = false
-        body.topAnchor.constraint(equalTo: view.topAnchor, constant: UI.padding).isActive = true
+        body.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: UI.padding).isActive = true
         body.bottomAnchor.constraint(
             equalTo: view.keyboardLayoutGuide.topAnchor,
             constant: -UI.padding
         ).isActive = true
         body.leftAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leftAnchor, constant: UI.padding).isActive = true
         body.rightAnchor.constraint(equalTo: view.safeAreaLayoutGuide.rightAnchor, constant: -UI.padding).isActive = true
+    }
+    
+    override func viewSafeAreaInsetsDidChange() {
+        super.viewSafeAreaInsetsDidChange()
+        content.heightAnchor.constraint(
+            equalTo: view.safeAreaLayoutGuide.heightAnchor,
+            constant: -UI.padding * 2
+        ).isActive = true
     }
 }